adicionada classe DAO para melhor refatoração e simplificação do código
