+++ MoviEveris/MoviEveris/Controller/DetalheFilmeViewController.swift
@@ -32,11 +32,11 @@ class DetalheFilmeViewController: UIViewController {
     
     func mostraFilme() {
         
-        guard let filme = filmeSelecionado else { return }
+        guard let filme = self.filmeSelecionado else { return }
         guard let idFilmeAtual = filme["id"] as? Int else { return }
         guard let imagemFilmeAtual = filme["poster"] as? UIImage else { return }
         
-        FilmeAPI().filmeDetalhes(idFilmeAtual) { (filme) in
+        FilmeDAO().filmeDetalhes(idFilmeAtual) { (filme) in
             var filmeAtual = filme[0]
             guard let tituloOriginal = filmeAtual["tituloOriginal"] as? String else { return }
             guard let titulo = filmeAtual["titulo"]  as? String else { return }
